Save work for phpipam_subnets (plural) data source pre-revert
Due to limitations in Terraform, interpolating lists with splat
operators and using interpolation functions does not really work as
intended. So, I'm moving this functionality to phpiapm_subnet, and
re-tooling it so that it will only return one subnet (as per the
singular data source). As a result, this code will be removed, so I'm
just saving the work before I do that.