Custom field support - nearly complete
Fields and tests are in, and nearly verified, for all resources that
support custom fields:

 * Addresses: resource and data source
 * Subnets: resource and data source (except with folders)
 * VLANs: resource only

The functionality allows you to both set custom fields, and search on
one in the data source via a regex.

Found a bug along the way with this with the API and folders, hence we
had to adjust the custom field search for subnets so it wasn't checking
folders. This might also be an issue with the subnet resource and we may
have to drop support for creating folders as a resource. Fetching on
with a data source should not be affected.